fixed what was broken in the xml
And ran Skatje's old validator script, which it turns out still finds about 140 more cases that need to be fixed where either indices don't line up in examples, or where there are old *PRO* and *trace* notations in the examples. I fixed a bunch-- the 140 is what is left. I have a spreadsheet.